feat: add daily-redefined cohort retention curves
Keeps the D1/D7/D30/D90 convention a founder already knows how to read, but
redefines 'retained'. For a product whose thesis is a contribution every day, an
account-based definition counts a user who invested once in week one and never
again as retained at D30 — exactly the churn the business needs to see.

Reports three definitions side by side because the gap between them is the
finding. Two are backward-looking and nested (invested on exactly day N; invested
in the prior 7 days); the third is forward-looking (invested any time on or after
day N) and is deliberately not a third step on the same scale — it crosses the
7-day window around D3, and that crossing is roughly where 'has an account'
stops resembling 'has a habit'.

Two things that keep the numbers honest: the denominator is the full user table,
including users who never transacted, since building it from the activity table
would drop the hardest churn; and cohorts without full exposure to a horizon are
excluded rather than divided by a partial denominator, which is what produces the
fake cliff in recent cohorts on most retention charts.

Pooling across cohorts is size-weighted so a small cohort cannot move the
headline as much as one ten times its size.

+++ src/analysis/survival_analysis.py
@@ -180,7 +180,9 @@ def build_gap_frame(streaks: pd.DataFrame, observation_end: pd.Timestamp) -> pd.
     days_remaining = (pd.to_datetime(observation_end) - breaks["streak_end"]).dt.days
 
     breaks["returned"] = observed_gap.notna().astype(int)
-    breaks["gap_duration"] = observed_gap.fillna(days_remaining).astype(float)
+    # Cast before filling: days_to_next_streak arrives as a nullable object column,
+    # and letting fillna do the downcast is deprecated in pandas 2.2.
+    breaks["gap_duration"] = observed_gap.astype("float64").fillna(days_remaining.astype("float64"))
 
     # A break on the final day has zero exposure and tells us nothing.
     breaks = breaks[breaks["gap_duration"] > 0]
